bashisms

+++ src/nativeMain/kotlin/borg/trikeshed/common/parser/readLines.kt
@@ -4,28 +4,7 @@ import borg.trikeshed.common.parser.simple.CharSeries
 import kotlinx.cinterop.*
 import platform.posix.*
 
-/** emulates the getline behavior instead of fgets
- *
-  *
- *     FILE *fp;
- *     char *line = NULL;
- *     size_t len = 0;
- *     ssize_t read;
- *     fp = fopen("/etc/motd", "r");
- *     if (fp == NULL)
- *         exit(1);
- *     while ((read = getline(&line, &len, fp)) != -1) {
- *         printf("Retrieved line of length %zu :\n", read);
- *         printf("%s", line);
- *     }
- *     if (ferror(fp)) {
- *         /* handle error */
- *     }
- *     free(line);
- *     fclose(fp);
- *     return 0;
- * }
- */
+/** lean on getline to read a file into a sequence of CharSeries */
 actual fun readLines(path: String): Sequence<CharSeries> = memScoped {
     return sequence {
         val fp = fopen(path, "r")
@@ -37,7 +16,7 @@ actual fun readLines(path: String): Sequence<CharSeries> = memScoped {
         val line: CPointerVarOf<CPointer<ByteVarOf<Byte>>> = alloc<CPointerVar<ByteVar>>()
         val len: ULongVarOf<size_t> = alloc<size_tVar>()
         len.value = 0u
-        var read: ssize_t = 0.toLong()
+        var read: ssize_t = 0L
 
         while (true) {
             read = getline(line.ptr, len.ptr, fp)
@@ -52,4 +31,3 @@ actual fun readLines(path: String): Sequence<CharSeries> = memScoped {
         }
     }
 }
-
